Explicit cast

+++ src/cpp/helper_fftw.cc
@@ -18,16 +18,16 @@ fftw_plan fft2_create_plan(vector<double> &input,
                            size_t nt)
 {
     // Define parameters
-    int rank = 2;                      // The rank of the fft (2 = fft2)
-    int n[2] = {ny, nx};               // Dimensions
-    int howmany = nt;                  // Number of fft2 to be computed
-    int *inembed = NULL;               // NULL is equivalent to passing n
-    int istride = 1;                   // Distance between two elements in the input
-    int idist = 2 * ny * (nx / 2 + 1); // Distance between k-th and (k+1)-th input elements
-    int *onembed = NULL;               // NULL is equivalent to passing n
-    int ostride = 1;                   // Distance between two elements in the output
-    int odist = ny * (nx / 2 + 1);     // Distance between k-th and (k+1)-th output elements
-    unsigned int flags = FFTW_MEASURE; // bitwise OR ('|') of zero or more planner flags (see http://www.fftw.org/fftw3.pdf)
+    int rank = 2;                             // The rank of the fft (2 = fft2)
+    int n[2] = {(int)ny, (int)nx};            // Dimensions
+    int howmany = (int)nt;                    // Number of fft2 to be computed
+    int *inembed = NULL;                      // NULL is equivalent to passing n
+    int istride = 1;                          // Distance between two elements in the input
+    int idist = (int)(2 * ny * (nx / 2 + 1)); // Distance between k-th and (k+1)-th input elements
+    int *onembed = NULL;                      // NULL is equivalent to passing n
+    int ostride = 1;                          // Distance between two elements in the output
+    int odist = (int)(ny * (nx / 2 + 1));     // Distance between k-th and (k+1)-th output elements
+    unsigned int flags = FFTW_MEASURE;        // bitwise OR ('|') of zero or more planner flags (see http://www.fftw.org/fftw3.pdf)
 
     // Create the fft2 plan
     fftw_plan plan = fftw_plan_many_dft_r2c(rank,
@@ -54,14 +54,14 @@ fftw_plan fft_create_plan(vector<double> &input,
 {
     // Define parameters
     int rank = 1;                      // The rank of the fft (1 = fft)
-    int n[1] = {nt};                   // Dimensions
-    int howmany = N;                   // Number of fft2 to be computed
+    int n[1] = {(int)nt};              // Dimensions
+    int howmany = (int)N;              // Number of fft2 to be computed
     int *inembed = NULL;               // NULL is equivalent to passing n
     int istride = 1;                   // Distance between two elements in the input
-    int idist = nt;                    // Distance between k-th and (k+1)-th input elements
+    int idist = (int)nt;               // Distance between k-th and (k+1)-th input elements
     int *onembed = NULL;               // NULL is equivalent to passing n
     int ostride = 1;                   // Distance between two elements in the output
-    int odist = nt;                    // Distance between k-th and (k+1)-th output elements
+    int odist = (int)nt;               // Distance between k-th and (k+1)-th output elements
     unsigned int flags = FFTW_MEASURE; // bitwise OR ('|') of zero or more planner flags (see http://www.fftw.org/fftw3.pdf)
 
     // Create the fft2 plan
@@ -90,14 +90,14 @@ fftw_plan ifft_create_plan(vector<double> &input,
 {
     // Define parameters
     int rank = 1;                      // The rank of the fft (1 = fft)
-    int n[1] = {nt};                   // Dimensions
-    int howmany = N;                   // Number of fft2 to be computed
+    int n[1] = {(int)nt};              // Dimensions
+    int howmany = (int)N;              // Number of fft2 to be computed
     int *inembed = NULL;               // NULL is equivalent to passing n
     int istride = 1;                   // Distance between two elements in the input
-    int idist = nt;                    // Distance between k-th and (k+1)-th input elements
+    int idist = (int)nt;               // Distance between k-th and (k+1)-th input elements
     int *onembed = NULL;               // NULL is equivalent to passing n
     int ostride = 1;                   // Distance between two elements in the output
-    int odist = nt;                    // Distance between k-th and (k+1)-th output elements
+    int odist = (int)nt;               // Distance between k-th and (k+1)-th output elements
     unsigned int flags = FFTW_MEASURE; // bitwise OR ('|') of zero or more planner flags (see http://www.fftw.org/fftw3.pdf)
 
     // Create the fft2 plan
